# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "OCPCTL is a self-service OpenShift cluster provisioning and management platform. This API provides endpoints for creating, managing, and destroying OpenShift clusters with automated lifecycle management, policy enforcement, and resource tracking.",
        "title": "OCPCTL API",
        "termsOfService": "http://github.com/tsanders-rh/ocpctl",
        "contact": {
            "name": "OCPCTL Support",
            "url": "http://github.com/tsanders-rh/ocpctl/issues"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/clusters/{id}/storage": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all storage groups linked to this cluster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Get cluster storage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_api.StorageGroupResponse"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - not cluster owner",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Cluster not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/clusters/{id}/storage/link": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Links persistent storage from another cluster to this cluster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Link storage to cluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Link storage request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.LinkStorageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.StorageGroupResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or validation error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - not cluster owner",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Cluster not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/clusters/{id}/storage/link/{group_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Removes the storage group link from this cluster",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Unlink storage from cluster",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Storage Group ID",
                        "name": "group_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - not cluster owner",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Cluster or storage link not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_tsanders-rh_ocpctl_pkg_types.ClusterStorageLinkRole": {
            "type": "string",
            "enum": [
                "source",
                "target",
                "shared"
            ],
            "x-enum-varnames": [
                "ClusterStorageLinkRoleSource",
                "ClusterStorageLinkRoleTarget",
                "ClusterStorageLinkRoleShared"
            ]
        },
...
        "github_com_tsanders-rh_ocpctl_pkg_types.StorageGroupStatus": {
            "type": "string",
            "enum": [
                "PROVISIONING",
                "READY",
                "FAILED",
                "DELETING"
            ],
            "x-enum-varnames": [
                "StorageGroupStatusProvisioning",
                "StorageGroupStatusReady",
                "StorageGroupStatusFailed",
                "StorageGroupStatusDeleting"
            ]
        },
...
        "internal_api.ClusterStorageLinkResponse": {
            "type": "object",
            "properties": {
                "cluster_id": {
                    "type": "string"
                },
                "cluster_name": {
                    "type": "string"
                },
                "linked_at": {
                    "type": "string"
                },
                "role": {
                    "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_pkg_types.ClusterStorageLinkRole"
                }
            }
        },
...
        "internal_api.LinkStorageRequest": {
            "type": "object",
            "required": [
                "target_cluster_id"
            ],
            "properties": {
                "target_cluster_id": {
                    "type": "string"
                }
            }
        },
...
        "internal_api.StorageGroupResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "efs_id": {
                    "type": "string"
                },
                "efs_security_group_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "linked_clusters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api.ClusterStorageLinkResponse"
                    }
                },
                "name": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "s3_bucket": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_tsanders-rh_ocpctl_pkg_types.StorageGroupStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT access token. Format: \"Bearer {token}\". Obtain from /api/v1/auth/login endpoint.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}